separate to new log_batch test file
